update to 2.3.1

+++ src/lib/helpers/http.ts
@@ -68,7 +68,9 @@ export class HttpClient {
                                 response.file = fs.createReadStream(file.path);
                                 response.file.on("end", () => {
                                     response.file.close();
-                                    fs.unlinkSync(response.file.path);
+                                    if (fs.existsSync(response.file.path)) {
+                                        fs.unlinkSync(response.file.path);
+                                    }
                                 });
                                 return resolve(response);
                             });
